UL-93: fix previous fix on html, now with jquery
now the document.write has been substituted with span tags of class
"widgetStrings" in the html. Inside widget.js a generic function resolves
the key to its value in the widgetStrings module. This mechanism has the
advantage of not needing to change the widget.js for subsequent changes in
the HTML or the widgetStrings.